move discussion proxy mapping from interactor
