+++ app/src/main/java/org/stepik/android/domain/discussion_proxy/model/DiscussionOrder.kt
@@ -1,0 +1,5 @@
+package org.stepik.android.domain.discussion_proxy.model
+
+enum class DiscussionOrder {
+    LAST_DISCUSSION, MOST_LIKED, MOST_ACTIVE, RECENT_ACTIVITY
+}
